refactor: make `Logo` a common component
