+++ src/components/Authentication/RegisterForm/style.css
@@ -23,11 +23,7 @@
 .auth-form .ant-form-item{
   margin-bottom: 5%;
 }
-/* LOGO */
-.home-layout-logo{
-  width: 175px;
-  height: 93px;
-}
+
 /* Title */
 .title-text{
   margin: 2% 0;
