+++ src/shared/components/Logo/index.jsx
