+++ src/shared/components/Logo/style.css
@@ -1,0 +1,4 @@
+.home-layout-logo{
+  width: 175px;
+  height: 93px;
+}
