+++ src/shared/components/index.js
@@ -1,2 +1,3 @@
 export { default as Loader } from './Spinner';
 export { default as UsersTable } from './Table';
+export { default as Logo } from './Logo';
